backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
transformers==4.35.2
torch==2.1.1
tensorflow==2.15.0
datasets==2.14.6
tokenizers==0.14.1
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
nltk==3.8.1
sentence-transformers==2.2.2
accelerate==0.24.1
evaluate==0.4.1
wandb==0.16.1
jupyter==1.0.0
notebook==7.0.6
rouge-score==0.1.2
sacrebleu==2.3.1
sacremoses==0.1.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
